Uses Real Access Point now
Instead of simulating an access point we now use a real one. The
access point is just a normal host that uses the access point
management module in ieee802.11 wireless interface. When a
subscriber is associated to a an access point it will broadcast
its subscription.

One remaining small problem is that because i use broadcasts to
reach the access point the ap will rebroadcast my message. This
could be solved to directly send the subscription to the ap but that
would make me dependent on ip layer which i dont want. It would
be better to somehow turn it off or just leave it open as it
wont matter anyway.